Simplify status code conversion
Instead of exposing a public staticmethod for converting status code to
status string, add a status_string property. Since status cannot be
invalid, it is not possible any more to convert an invalid value, so we
don't need the error handler.

Also change the private helper to convert status string to status code
to an instance method, since has access to the _STATUS_CONV table.